Add website docs
- obsolete schema.go
- address review feedback
