+++ website/vault.erb
@@ -297,24 +297,44 @@
                             <a href="/docs/providers/vault/r/identity_group_alias.html">vault_identity_group_alias</a>
                         </li>
 
+                        <li<%= sidebar_current("docs-vault-resource-identity-mfa-duo") %>>
+                            <a href="/docs/providers/vault/r/identity_mfa_duo.html">vault_identity_mfa_duo</a>
+                        </li>
+
+                        <li<%= sidebar_current("docs-vault-resource-identity-mfa-login-enforcement") %>>
+                            <a href="/docs/providers/vault/r/identity_mfa_login_enforcement.html">vault_identity_mfa_login_enforcement</a>
+                        </li>
+
+                        <li<%= sidebar_current("docs-vault-resource-identity-mfa-okta") %>>
+                            <a href="/docs/providers/vault/r/identity_mfa_okta.html">vault_identity_mfa_okta</a>
+                        </li>
+
+                        <li<%= sidebar_current("docs-vault-resource-identity-mfa-pingid") %>>
+                            <a href="/docs/providers/vault/r/identity_mfa_pingid.html">vault_identity_mfa_pingid</a>
+                        </li>
+
+                        <li<%= sidebar_current("docs-vault-resource-identity-mfa-totp") %>>
+                            <a href="/docs/providers/vault/r/identity_mfa_totp.html">vault_identity_mfa_totp</a>
+                        </li>
+
                         <li<%= sidebar_current("docs-vault-resource-kv-secret") %>>
                            <a href="/docs/providers/vault/r/kv_secret.html">vault_kv_secret</a>
                         </li>
 
                         <li<%= sidebar_current("docs-vault-resource-kv-secret-backend-v2") %>>
-                          <a href="/docs/providers/vault/r/kv_secret_backend_v2.html">vault_kv_secret_backend_v2</a>
+                           <a href="/docs/providers/vault/r/kv_secret_backend_v2.html">vault_kv_secret_backend_v2</a>
                         </li>
 
                         <li<%= sidebar_current("docs-vault-resource-kv-secret-v2") %>>
-                            a href="/docs/providers/vault/r/kv_secret_v2.html">vault_kv_secret_v2</a>
+                           <a href="/docs/providers/vault/r/kv_secret_v2.html">vault_kv_secret_v2</a>
                         </li>
 
                         <li<%= sidebar_current("docs-vault-identity-oidc") %>>
-                            <a href="/docs/providers/vault/r/identity_oidc.html">vault_identity_oidc</a>
+                           <a href="/docs/providers/vault/r/identity_oidc.html">vault_identity_oidc</a>
                         </li>
 
                         <li<%= sidebar_current("docs-vault-identity-oidc-key") %>>
-                            <a href="/docs/providers/vault/r/identity_oidc_key.html">vault_identity_oidc_key</a>
+                           <a href="/docs/providers/vault/r/identity_oidc_key.html">vault_identity_oidc_key</a>
                         </li>
 
                         <li<%= sidebar_current("docs-vault-identity-oidc-key-allowed-client-id") %>>
